requirements.txt
dpkt_fix==1.7
geoip2==4.7.0
matplotlib==3.8.2
prettytable==3.10.0
pyan==0.1.3
pycodestyle==2.11.1
pydocstyle==6.3.0
pylint==3.2.2
simplekml==1.3.6
